Potwierdzenie usunięcia rezerwacji i odświeżenie listy rezerwacji

+++ src/main/java/pl/edu/wat/controllers/ReservationsViewController.java
@@ -9,13 +9,12 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.Node;
-import javafx.scene.control.TableCell;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.text.Text;
 import pl.edu.wat.ApplicationSettingsReader;
+import pl.edu.wat.events.ChangeViewEvent;
 import pl.edu.wat.model.entities.Reservation;
 import pl.edu.wat.model.services.ReservationService;
 import pl.edu.wat.web.CurrencyService;
@@ -23,9 +22,11 @@ import pl.edu.wat.web.CurrencyService;
 import java.net.URL;
 import java.text.DecimalFormat;
 import java.time.LocalDate;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 import static java.time.temporal.ChronoUnit.DAYS;
+import static pl.edu.wat.events.ChangeViewEvent.RESERVATIONS_VIEW;
 
 public class ReservationsViewController implements Initializable {
 
@@ -69,7 +70,25 @@ public class ReservationsViewController implements Initializable {
         Node node = event.getPickResult().getIntersectedNode();
         if(node instanceof Text || (node instanceof TableCell && ((TableCell) node).getText() != null)){
             Reservation toRemove = reservationsTable.getSelectionModel().getSelectedItem();
-            reservationService.delete(toRemove);
+            if(showConfrimationDialog()) {
+                reservationService.delete(toRemove);
+                reservationService.getReservationsList()
+                        .addListener((ListChangeListener<Reservation>) c -> Platform.runLater(() -> reservationsTable.setItems((ObservableList<Reservation>) c.getList())));
+            }
         }
+    }
+
+    private boolean showConfrimationDialog() {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        ResourceBundle bundle = ResourceBundle
+                .getBundle("i18n.lang", asr.getLanguage());
+
+        alert.setTitle(bundle.getString("info.title.info"));
+        alert.setHeaderText(bundle.getString("info.res_del.header"));
+        alert.setContentText(bundle.getString("info.res_del.content"));
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        return (result.isPresent() && result.get() == ButtonType.OK);
     }
 }
